requirements.txt
numpy
opencv-python
torch
torchvision
Pillow
matplotlib
ultralytics
keyboard